добавлена логика для получения уроков

+++ app/src/main/java/stas/batura/schedulermonth/repository/room/Period.kt
@@ -18,7 +18,7 @@ data class Period(
 //    var lessonIsComplete: Int = 0,
 
     @ColumnInfo(name = "section_id")
-    var sectionId : Int = -1
+    var sectionId : Long = -1
 
 ) {
     @PrimaryKey(autoGenerate = true)
